kkg-area-detection-develop/pyproject.toml
[project]
name = "kkg-area-detection"
version = "0.1.0"
description = "Library for area detection in images"
authors = [
    {name = "Kohei Ichikawa", email = "Kohei_Ichikawa@kk-generation.com"},
    {name = "Kenichiro Ando"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.3",
    "scipy>=1.10.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "opencv-contrib-python>=4.8.0",
    "azure-ai-formrecognizer>=3.3.3",
    "azure-core>=1.29.5",
    "svg.path>=2.2.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.66.1",
    "python-dotenv>=1.1.0",
    "boto3>=1.34.69",
    "typing-extensions>=4.7.0",
    "shapely>=2.1.1",
    "ocrd-fork-pylsd>=0.0.8",
    "scikit-image>=0.19.0",
    "sknw>=0.14.2",
    "rdp>=0.8",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "autopep8",
    "pytest",
    "flake8==6.1.0",
    "flake8-pyproject",
    "flake8-commas==2.1.0",
    "flake8-isort==6.1.0",
    "flake8-quotes==3.3.2",
    "mypy",
    "black",
    "isort",
    "ruff"
]

notebook = [
    "ipykernel>=6.25.0",
    "ipywidgets>=8.1.5",
    "ipython==8.16.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["kkg_area_detection*"]
exclude = ["tests*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.flake8]
max-line-length = 120
ignore = ["C812", "Q000", "W503", "W504"]
exclude = ["tests*", "sandbox*"]
